feat: hunspell and morfologik spellchecker modules
- Add morfologik spellcheck module
- Rename hunspell-jmyspell to hunspell module
- Drop jmyspell fallback
- ISpellChecker#initialize return boolean
