+++ build.gradle
@@ -257,7 +257,6 @@ dependencies {
         implementation(libs.omegat.htmlparser)
         implementation(libs.omegat.gnudiff4j)
         implementation(libs.omegat.mnemonics)
-        implementation(libs.omegat.jmyspell.core)
 
         // LanguageTool
         implementation(libs.languagetool.core) {
