+++ gradle/libs.versions.toml
@@ -43,7 +43,6 @@ dsl4j = "1.0.0"
 stardict4j = "1.1.0"
 juniversalchardet = "2.4.0"
 mnemonics = "1.2"
-jmyspell = "1.0.0-beta-2"
 hunspell = "2.1.2"
 xjc = "2.3.4"
 jna = "5.13.0"
@@ -51,6 +50,7 @@ jfa = "1.2.0"
 tipoftheday = "0.4.4"
 flatlaf="3.4.1"
 assertj_swing_junit = "4.0.0-beta-1"
+morfologik = "2.1.9"
 
 [libraries]
 slf4j-api = {group = "org.slf4j", name = "slf4j-api",  version.ref = "slf4j"}
@@ -128,7 +128,6 @@ dsl4j = {group = "tokyo.northside", name = "dsl4j", version.ref = "dsl4j"}
 stardict4j = {group = "tokyo.northside", name = "stardict4j", version.ref = "stardict4j"}
 juniversal-chardet = {group = "com.github.albfernandez", name = "juniversalchardet", version.ref = "juniversalchardet"}
 omegat-mnemonics = {group = "org.omegat", name = "lib-mnemonics", version.ref = "mnemonics"}
-omegat-jmyspell-core = {group = "org.omegat", name = "jmyspell-core", version.ref = "jmyspell"}
 dumont-hunspell = {group = "com.gitlab.dumonts", name = "hunspell", version.ref = "hunspell"}
 jaxb-xjc = {group = "org.grassfish.jaxb", name = "jaxb-xjc", version.ref = "xjc"}
 tipoftheday = {group = "tokyo.northside", name = "tipoftheday", version.ref = "tipoftheday"}
@@ -137,6 +136,8 @@ jfa = {group = "de.jangassen", name = "jfa", version.ref = "jfa"}
 flatlaf = {group = "com.formdev", name = "flatlaf", version.ref = "flatlaf"}
 language-detector = {group = "org.omegat", name = "language-detector", version.ref = "languagedetector"}
 assertj_swing_junit = {group = "tokyo.northside", name = "assertj-swing-junit", version.ref = "assertj_swing_junit"}
+morfologik = { group = "org.carrot2", name = "morfologik-stemming", version.ref = "morfologik" }
+morfologik-speller = { group = "org.carrot2", name = "morfologik-speller", version.ref = "morfologik" }
 
 [bundles]
 groovy = ["groovy-jsr223", "groovy-dateutil", "groovy-json", "groovy-xml", "groovy-swing", "groovy-templates", "ivy"]
@@ -150,6 +151,7 @@ lucene = ["lucene-core", "lucene-analyzers-common", "lucene-analyzers-kuromoji",
 jgit = ["jgit", "jgit-agent", "jgit-http", "jgit-ssh"]
 dictionary = ["trie4j", "dsl4j", "stardict4j", "jsoup"]
 xmlunit = ["xmlunit-core", "xmlunit-assertj", "assertj"]
+morfologik = ['morfologik', 'morfologik-speller']
 
 [plugins]
 spotbugs = {id = "com.github.spotbugs", version = "6.0.15"}
