--- spellchecker/hunspell-jmyspell/src/main/resources/org/omegat/spellchecker/Bundle.properties
@@ -1,56 +1,0 @@
-#
-#  OmegaT - Computer Assisted Translation (CAT) tool
-#           with fuzzy matching, translation memory, keyword search,
-#           glossaries, and translation leveraging into updated projects.
-#
-#  Copyright (C) 2023 Hiroshi Miura
-#                Home page: https://www.omegat.org/
-#                Support center: https://omegat.org/support
-#
-#  This file is part of OmegaT.
-#
-#  OmegaT is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#
-#  OmegaT is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-ERROR_TITLE=Error
-
-SPELLCHECKER_LANGUAGE_NOT_FOUND=No spell checker found for language {0}
-SPELLCHECKER_HUNSPELL_INITIALIZED=Initialized LanguageTool Hunspell spellchecker for language '{0}' dictionary {1}
-SPELLCHECKER_HUNSPELL_EXCEPTION=Error loading hunspell: {0}
-SPELLCHECKER_JMYSPELL_INITIALIZED=Initialized JMySpell spell checker for language '{0}' dictionary {1}
-SPELLCHECKER_JMYSPELL_EXCEPTION=Error loading jmyspell: {0}
-SPELLCHECKER_DICTIONARY_NOT_FILE=Spelling dictionary exists but is not a file: {0}
-SPELLCHECKER_DICTIONARY_NOT_READ=Can't read spelling dictionary: {0}
-SPELLCHECKER_DICTIONARY_EMPTY=Spelling dictionary appears to be empty: {0}
-
-
-GUI_SPELLCHECKER_TITLE=Spellchecker Setup
-PREFS_TITLE_SPELLCHECKER=Spellchecker
-GUI_SPELLCHECKER_AUTOSPELLCHECKCHECKBOX=Spellcheck the tr&anslation
-GUI_SPELLCHECKER_AVAILABLE_LABEL=Available languages:
-GUI_SPELLCHECKER_DICTIONARYLABEL=Spellchecking dictionaries location:
-GUI_SPELLCHECKER_FILE_CHOOSER_TITLE=Dictionary File Folder
-GUI_SPELLCHECKER_DIRECTORYCHOOSERBUTTON=&Browse...
-GUI_SPELLCHECKER_INSTALLBUTTON=&Install new language...
-GUI_SPELLCHECKER_UNINSTALLBUTTON=D&elete
-GUI_SPELLCHECKER_URL_LABEL=URL of downloadable spellchecking dictionaries:
-GUI_SPELLCHECKER_UNINSTALL_CURRENT=You are deleting the spellchecking dictionary that your project currently uses. Proceed?
-GUI_SPELLCHECKER_UNINSTALL_CURRENT_TITLE=Confirm deletion
-GUI_SPELLCHECKER_UNINSTALL_UNABLE=Unable to delete the selected dictionary
-GUI_SPELLCHECKER_UNINSTALL_UNABLE_TITLE=Deletion failed
-GUI_SPELLCHECKER_INSTALL_UNABLE_TITLE=Install failed
-GUI_SPELLCHECKER_INSTALL_UNABLE=The specified spellchecking dictionary folder cannot be used.
-GUI_SPELLCHECKER_DIR_NOT_PRESENT_TITLE=Folder does not exist
-GUI_SPELLCHECKER_DIR_NOT_PRESENT=The specified spellchecking dictionary folder does not exist. Do you want to create it?
-GUI_SPELLCHECKER_COULD_NOT_CREATE_DIR=Could not create the specified folder.
-GUI_SPELLCHECKER_ERROR_ON_INSTALL=An error occurred while downloading the spellchecking dictionary. Try again later.
