--- spellchecker/hunspell-jmyspell/src/main/resources/org/omegat/spellchecker/Bundle_de.properties
@@ -1,81 +1,0 @@
-#/**************************************************************************
-# OmegaT - Computer Assisted Translation (CAT) tool
-#          with fuzzy matching, translation memory, keyword search,
-#          glossaries, and translation leveraging into updated projects.
-#
-# Copyright (C) 2000-2006 Keith Godfrey, Maxym Mykhalchuk, and Henry Pijffers
-#               2007 Didi- Briel, Zoltan Bartko, Martin Fleurke
-#               2008 Andrzej Sawula, Alex Buloichik, Didier Briel, Martin Fleurke
-#               2009 Didier Briel, Alex Buloichik, Martin Fleurke
-#               2010 Didier Briel, Wildrich Fourie, Martin Fleurke,
-#                    Antonio Vilei, Arno Peters, Alex Buloichik, Piotr Kulik
-#               2011 Didier Briel, Briac Pilpre
-#               2012 Martin Fleurke, Wildrich Fourie, Didier Briel, Guido Leenders, Jean-Christophe Helary,
-#                    Alex Buloichik
-#               2013 Aaron Madlon-Kay, Zoltan Bartko, Didier Briel
-#               2014 Aaron Madlon-Kay, Didier Briel, Piotr Kulik, Alex Buloichik, Briac Pilpre
-#               2015 Yu Tang, Aaron Madlon-Kay, Didier Briel, Hiroshi Miura
-#               2016 Alex Buloichik, Aaron Madlon-Kay, Didier Briel, Briac Pilpre
-#               2017 Didier Briel
-#               2020-2022 Hiroshi Miura
-#               Home page: https://www.omegat.org/
-#               Support centre: https://omegat.org/support
-#
-# This file is part of OmegaT.
-#
-# OmegaT is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# OmegaT is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#**************************************************************************/
-#
-#  Resources file for OmegaT
-#
-#  @author Keith Godfrey
-#  @author Maxym Mykhalchuk
-#  @author Henry Pijffers
-#  @author Didier Briel
-#  @author Jean-Christophe Helary
-#  @author Zoltan Bartko - bartkozoltan@bartkozoltan.com
-#  @author Martin Fleurke
-#  @author Andrzej Sawula
-#  @author Alex Buloichik
-#  @author Wildrich Fourie
-#  @author Antonio Vilei
-#  @author Briac Pilpre
-#  @author Aaron Madlon-Kay
-#  @author Piotr Kulik
-#  @author Briac Pilpre
-#
-#  @version 4.1
-#
-
-GUI_SPELLCHECKER_TITLE=Rechtschreibprüfung
-PREFS_TITLE_SPELLCHECKER=Rechtschreibprüfung
-GUI_SPELLCHECKER_AUTOSPELLCHECKCHECKBOX=Rechtschreibung der Übersetzung &prüfen
-GUI_SPELLCHECKER_AVAILABLE_LABEL=Verfügbare Sprachen:
-GUI_SPELLCHECKER_DICTIONARYLABEL=Speicherort der Rechtschreibwörterbücher:
-GUI_SPELLCHECKER_FILE_CHOOSER_TITLE=Ordner mit Wörterbuchdateien
-GUI_SPELLCHECKER_DIRECTORYCHOOSERBUTTON=Du&rchsuchen...
-GUI_SPELLCHECKER_INSTALLBUTTON=Neue Sprache &installieren...
-GUI_SPELLCHECKER_UNINSTALLBUTTON=Lösch&en
-GUI_SPELLCHECKER_URL_LABEL=URL von herunterladbaren Rechtschreibwörterbüchern:
-GUI_SPELLCHECKER_UNINSTALL_CURRENT=Sie löschen gerade das Wörterbuch, das Ihr Projekt zur Rechtschreibprüfung verwendet. Fortsetzen?
-GUI_SPELLCHECKER_UNINSTALL_CURRENT_TITLE=Löschung bestätigen
-GUI_SPELLCHECKER_UNINSTALL_UNABLE=Das ausgewählte Wörterbuch kann nicht gelöscht werden
-GUI_SPELLCHECKER_UNINSTALL_UNABLE_TITLE=Löschung fehlgeschlagen
-GUI_SPELLCHECKER_INSTALL_UNABLE_TITLE=Installation fehlgeschlagen
-GUI_SPELLCHECKER_INSTALL_UNABLE=Der angegebene Rechtschreibwörterbuch-Ordner kann nicht verwendet werden.
-GUI_SPELLCHECKER_DIR_NOT_PRESENT_TITLE=Ordner existiert nicht
-GUI_SPELLCHECKER_DIR_NOT_PRESENT=Der angegebene Rechtschreibwörterbuch-Ordner existiert nicht. Möchten Sie ihn erstellen?
-GUI_SPELLCHECKER_COULD_NOT_CREATE_DIR=Der angegebene Ordner konnte nicht erstellt werden.
-GUI_SPELLCHECKER_ERROR_ON_INSTALL=Beim Herunterladen des Rechtschreibwörterbuchs ist ein Fehler aufgetreten. Versuchen Sie es später noch einmal.
-
