--- spellchecker/hunspell-jmyspell/src/main/resources/org/omegat/spellchecker/Bundle_es.properties
@@ -1,79 +1,0 @@
-#/**************************************************************************
-# OmegaT - Computer Assisted Translation (CAT) tool
-#          with fuzzy matching, translation memory, keyword search,
-#          glossaries, and translation leveraging into updated projects.
-#
-# Copyright (C) 2000-2006 Keith Godfrey, Maxym Mykhalchuk, and Henry Pijffers
-#               2007 Didier Briel, Zoltan Bartko, Martin Fleurke
-#               2008 Andrzej Sawula, Alex Buloichik, Didier Briel, Martin Fleurke
-#               2009 Didier Briel, Alex Buloichik, Martin Fleurke
-#               2010 Didier Briel, Wildrich Fourie, Martin Fleurke,
-#                    Antonio Vilei, Arno Peters, Alex Buloichik, Piotr Kulik
-#               2011 Didier Briel, Briac Pilpre
-#               2012 Martin Fleurke, Wildrich Fourie, Didier Briel, Guido Leenders, Jean-Christophe Helary,
-#                    Alex Buloichik
-#               2013 Aaron Madlon-Kay, Zoltan Bartko, Didier Briel
-#               2014 Aaron Madlon-Kay, Didier Briel, Piotr Kulik, Alex Buloichik, Briac Pilpre
-#               2015 Yu Tang, Aaron Madlon-Kay, Didier Briel, Hiroshi Miura
-#               2016 Alex Buloichik, Aaron Madlon-Kay, Didier Briel, Briac Pilpre
-#               2017 Didier Briel
-#               Home page: https://www.omegat.org/
-#               Support center: https://omegat.org/support
-#
-# This file is part of OmegaT.
-#
-# OmegaT is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# OmegaT is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#**************************************************************************/
-#
-#  Resources file for OmegaT
-#
-#  @author Keith Godfrey
-#  @author Maxym Mykhalchuk
-#  @author Henry Pijffers
-#  @author Didier Briel
-#  @author Jean-Christophe Helary
-#  @author Zoltan Bartko - bartkozoltan@bartkozoltan.com
-#  @author Martin Fleurke
-#  @author Andrzej Sawula
-#  @author Alex Buloichik
-#  @author Wildrich Fourie
-#  @author Antonio Vilei
-#  @author Briac Pilpre
-#  @author Aaron Madlon-Kay
-#  @author Piotr Kulik
-#  @author Briac Pilpre
-#
-
-
-GUI_SPELLCHECKER_TITLE=Configuración del corrector ortográfico
-PREFS_TITLE_SPELLCHECKER=Corrector ortográfico
-GUI_SPELLCHECKER_AUTOSPELLCHECKCHECKBOX=Comprobar la ortografía &automáticamente
-GUI_SPELLCHECKER_AVAILABLE_LABEL=Diccionarios ya instalados:
-GUI_SPELLCHECKER_DICTIONARYLABEL=Carpeta de archivos de diccionarios:
-GUI_SPELLCHECKER_FILE_CHOOSER_TITLE=Carpeta de archivos de diccionarios
-GUI_SPELLCHECKER_DIRECTORYCHOOSERBUTTON=&Explorar...
-GUI_SPELLCHECKER_INSTALLBUTTON=&Instalar nuevo diccionario...
-GUI_SPELLCHECKER_UNINSTALLBUTTON=Quita&r
-GUI_SPELLCHECKER_URL_LABEL=URL de diccionarios en línea:
-GUI_SPELLCHECKER_UNINSTALL_CURRENT=Está desinstalando el diccionario de corrección ortográfica que está utilizando su proyecto. ¿Desea continuar?
-GUI_SPELLCHECKER_UNINSTALL_CURRENT_TITLE=Confirmar desinstalación
-GUI_SPELLCHECKER_UNINSTALL_UNABLE=No se puede desinstalar el diccionario seleccionado
-GUI_SPELLCHECKER_UNINSTALL_UNABLE_TITLE=Falló la desinstalación
-GUI_SPELLCHECKER_INSTALL_UNABLE_TITLE=No se puede instalar
-GUI_SPELLCHECKER_INSTALL_UNABLE=No se puede utilizar la carpeta de diccionarios especificada.
-GUI_SPELLCHECKER_DIR_NOT_PRESENT_TITLE=La carpeta no existe
-GUI_SPELLCHECKER_DIR_NOT_PRESENT=La carpeta de diccionarios especificada no existe. ¿Quiere crearla?
-GUI_SPELLCHECKER_COULD_NOT_CREATE_DIR=No se ha podido crear la carpeta especificada.
-GUI_SPELLCHECKER_ERROR_ON_INSTALL=Error al descargar el diccionario. Inténtelo de nuevo.
-
