--- spellchecker/hunspell-jmyspell/src/main/resources/org/omegat/spellchecker/Bundle_fr.properties
@@ -1,60 +1,0 @@
-#
-#  OmegaT - Computer Assisted Translation (CAT) tool
-#           with fuzzy matching, translation memory, keyword search,
-#           glossaries, and translation leveraging into updated projects.
-#
-# Copyright (C) 2000-2006 Keith Godfrey, Maxym Mykhalchuk, and Henry Pijffers
-#               2007 Didi- Briel, Zoltan Bartko, Martin Fleurke
-#               2008 Andrzej Sawula, Alex Buloichik, Didier Briel, Martin Fleurke
-#               2009 Didier Briel, Alex Buloichik, Martin Fleurke
-#               2010 Didier Briel, Wildrich Fourie, Martin Fleurke,
-#                    Antonio Vilei, Arno Peters, Alex Buloichik, Piotr Kulik
-#               2011 Didier Briel, Briac Pilpre
-#               2012 Martin Fleurke, Wildrich Fourie, Didier Briel, Guido Leenders, Jean-Christophe Helary,
-#                    Alex Buloichik
-#               2013 Aaron Madlon-Kay, Zoltan Bartko, Didier Briel
-#               2014 Aaron Madlon-Kay, Didier Briel, Piotr Kulik, Alex Buloichik, Briac Pilpre
-#               2015 Yu Tang, Aaron Madlon-Kay, Didier Briel, Hiroshi Miura
-#               2016 Alex Buloichik, Aaron Madlon-Kay, Didier Briel, Briac Pilpre
-#               2017 Didier Briel
-#               2020-2022 Hiroshi Miura
-#               Home page: https://www.omegat.org/
-#               Support centre: https://omegat.org/support
-#
-#  This file is part of OmegaT.
-#
-#  OmegaT is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#
-#  OmegaT is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-
-GUI_SPELLCHECKER_TITLE=Configuration du vérificateur orthographique
-PREFS_TITLE_SPELLCHECKER=Vérificateur orthographique
-GUI_SPELLCHECKER_AUTOSPELLCHECKCHECKBOX=Vérifier l’orthographe de la tr&aduction
-GUI_SPELLCHECKER_AVAILABLE_LABEL=Langues disponibles :
-GUI_SPELLCHECKER_DICTIONARYLABEL=Emplacement des dictionnaires de vérification orthographique.
-GUI_SPELLCHECKER_FILE_CHOOSER_TITLE=Dossier des fichiers dictionnaire
-GUI_SPELLCHECKER_DIRECTORYCHOOSERBUTTON=Sél&ectionner
-GUI_SPELLCHECKER_INSTALLBUTTON=&Installer une nouvelle langue…
-GUI_SPELLCHECKER_UNINSTALLBUTTON=&Effacer
-GUI_SPELLCHECKER_URL_LABEL=URL des dictionnaires orthographiques téléchargeables :
-GUI_SPELLCHECKER_UNINSTALL_CURRENT=Vous allez effacer le dictionnaire orthographique utilisé actuellement par votre projet. Voulez-vous continuer ?
-GUI_SPELLCHECKER_UNINSTALL_CURRENT_TITLE=Confirmer l’effacement
-GUI_SPELLCHECKER_UNINSTALL_UNABLE=Impossible d’effacer le dictionnaire sélectionné
-GUI_SPELLCHECKER_UNINSTALL_UNABLE_TITLE=L’effacement a échoué
-GUI_SPELLCHECKER_INSTALL_UNABLE_TITLE=L’installation a échoué
-GUI_SPELLCHECKER_INSTALL_UNABLE=Le dossier de dictionnaires orthographiques spécifié ne peut pas être utilisé.
-GUI_SPELLCHECKER_DIR_NOT_PRESENT_TITLE=Le dossier n’existe pas
-GUI_SPELLCHECKER_DIR_NOT_PRESENT=Le dossier de dictionnaires orthographiques spécifié n’existe pas. Voulez-vous le créer ?
-GUI_SPELLCHECKER_COULD_NOT_CREATE_DIR=Impossible de créer le dossier spécifié.
-GUI_SPELLCHECKER_ERROR_ON_INSTALL=Une erreur s’est produite durant le téléchargement du dictionnaire orthographique. Réessayez plus tard.
-
