--- spellchecker/hunspell-jmyspell/src/main/resources/org/omegat/spellchecker/Bundle_ja.properties
@@ -1,56 +1,0 @@
-#
-#  OmegaT - Computer Assisted Translation (CAT) tool
-#           with fuzzy matching, translation memory, keyword search,
-#           glossaries, and translation leveraging into updated projects.
-#
-#  Copyright (C) 2023 Hiroshi Miura
-#                Home page: https://www.omegat.org/
-#                Support center: https://omegat.org/support
-#
-#  This file is part of OmegaT.
-#
-#  OmegaT is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#
-#  OmegaT is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-ERROR_TITLE=\u30A8\u30E9\u30FC
-
-SPELLCHECKER_LANGUAGE_NOT_FOUND=\u8A00\u8A9E{0}\u306E\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AB\u30FC\u304C\u898B\u3064\u304B\u308A\u307E\u305B\u3093\u3067\u3057\u305F
-SPELLCHECKER_HUNSPELL_INITIALIZED=\u8A00\u8A9E '{0}' \u7528\u306E LanguageTool Hunspell \u7DB4\u3044\u78BA\u8A8D\u7528\u8F9E\u66F8\u3092\u521D\u671F\u5316 {1} 
-SPELLCHECKER_HUNSPELL_EXCEPTION=hunspell\u306E\u8AAD\u307F\u8FBC\u307F\u30A8\u30E9\u30FC:{0}
-SPELLCHECKER_JMYSPELL_INITIALIZED=\u8A00\u8A9E\u306E\u300C{0}\u300D\u8F9E\u66F8{1}\u7528\u306EJMySpell\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AB\u30FC\u3092\u521D\u671F\u5316\u3057\u307E\u3057\u305F
-SPELLCHECKER_JMYSPELL_EXCEPTION=jmyspell\u306E\u8AAD\u307F\u8FBC\u307F\u4E2D\u306B\u30A8\u30E9\u30FC\u304C\u767A\u751F\u3057\u307E\u3057\u305F:{0}
-SPELLCHECKER_DICTIONARY_NOT_FILE=\u7DB4\u308A\u78BA\u8A8D\u7528\u8F9E\u66F8\u306F\u5B58\u5728\u3057\u307E\u3059\u304C\u3001\u30D5\u30A1\u30A4\u30EB\u3067\u306F\u3042\u308A\u307E\u305B\u3093: {0}
-SPELLCHECKER_DICTIONARY_NOT_READ=\u7DB4\u308A\u78BA\u8A8D\u7528\u8F9E\u66F8 {0} \u304C\u8AAD\u307F\u53D6\u308C\u307E\u305B\u3093
-SPELLCHECKER_DICTIONARY_EMPTY=\u7DB4\u308A\u78BA\u8A8D\u7528\u8F9E\u66F8\u304C\u7A7A\u306E\u3088\u3046\u3067\u3059\u3002
-
-
-GUI_SPELLCHECKER_TITLE=\u7DB4\u308A\u78BA\u8A8D\u6A5F\u80FD\u306E\u8A2D\u5B9A
-PREFS_TITLE_SPELLCHECKER=\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AB\u30FC
-GUI_SPELLCHECKER_AUTOSPELLCHECKCHECKBOX=\u8A33\u6587\u3092\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AF(&A)
-GUI_SPELLCHECKER_AVAILABLE_LABEL=\u5229\u7528\u53EF\u80FD\u306A\u8A00\u8A9E:
-GUI_SPELLCHECKER_DICTIONARYLABEL=\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AF\u8F9E\u66F8\u306E\u5834\u6240:
-GUI_SPELLCHECKER_FILE_CHOOSER_TITLE=\u8F9E\u66F8\u30D5\u30A1\u30A4\u30EB\u30D5\u30A9\u30EB\u30C0\u30FC
-GUI_SPELLCHECKER_DIRECTORYCHOOSERBUTTON=\u53C2\u7167(&B)...
-GUI_SPELLCHECKER_INSTALLBUTTON=\u65B0\u3057\u3044\u8A00\u8A9E\u3092\u30A4\u30F3\u30B9\u30C8\u30FC\u30EB(&I)
-GUI_SPELLCHECKER_UNINSTALLBUTTON=\u524A\u9664(&E)
-GUI_SPELLCHECKER_URL_LABEL=\u30C0\u30A6\u30F3\u30ED\u30FC\u30C9\u53EF\u80FD\u306A\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AF\u8F9E\u66F8\u306EURL:
-GUI_SPELLCHECKER_UNINSTALL_CURRENT=\u73FE\u5728\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u304C\u4F7F\u7528\u3057\u3066\u3044\u308B\u8F9E\u66F8\u3092\u524A\u9664\u3057\u3088\u3046\u3068\u3057\u3066\u3044\u307E\u3059\u3002\u7D9A\u884C\u3057\u307E\u3059\u304B\uFF1F
-GUI_SPELLCHECKER_UNINSTALL_CURRENT_TITLE=\u524A\u9664\u306E\u78BA\u8A8D
-GUI_SPELLCHECKER_UNINSTALL_UNABLE=\u9078\u629E\u3057\u305F\u8F9E\u66F8\u3092\u524A\u9664\u3067\u304D\u307E\u305B\u3093\u3002
-GUI_SPELLCHECKER_UNINSTALL_UNABLE_TITLE=\u524A\u9664\u306B\u5931\u6557\u3057\u307E\u3057\u305F
-GUI_SPELLCHECKER_INSTALL_UNABLE_TITLE=\u30A4\u30F3\u30B9\u30C8\u30FC\u30EB\u306B\u5931\u6557\u3057\u307E\u3057\u305F
-GUI_SPELLCHECKER_INSTALL_UNABLE=\u6307\u5B9A\u3055\u308C\u305F\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AB\u30FC\u8F9E\u66F8\u30D5\u30A9\u30EB\u30C0\u30FC\u3092\u4F7F\u7528\u3067\u304D\u307E\u305B\u3093\u3002
-GUI_SPELLCHECKER_DIR_NOT_PRESENT_TITLE=\u30D5\u30A9\u30EB\u30C0\u30FC\u304C\u898B\u3064\u304B\u308A\u307E\u305B\u3093
-GUI_SPELLCHECKER_DIR_NOT_PRESENT=\u6307\u5B9A\u3055\u308C\u305F\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AB\u30FC\u8F9E\u66F8\u30D5\u30A9\u30EB\u30C0\u30FC\u304C\u898B\u3064\u304B\u308A\u307E\u305B\u3093\u3067\u3057\u305F\u3002\u30D5\u30A9\u30EB\u30C0\u30FC\u3092\u4F5C\u6210\u3057\u307E\u3059\u304B\uFF1F
-GUI_SPELLCHECKER_COULD_NOT_CREATE_DIR=\u6307\u5B9A\u3055\u308C\u305F\u30D5\u30A9\u30EB\u30C0\u30FC\u3092\u4F5C\u6210\u3067\u304D\u307E\u305B\u3093\u3067\u3057\u305F\u3002
-GUI_SPELLCHECKER_ERROR_ON_INSTALL=\u30B9\u30DA\u30EB\u30C1\u30A7\u30C3\u30AB\u30FC\u8F9E\u66F8\u306E\u30C0\u30A6\u30F3\u30ED\u30FC\u30C9\u4E2D\u306B\u30A8\u30E9\u30FC\u304C\u767A\u751F\u3057\u307E\u3057\u305F\u3002\u5B9F\u884C\u3057\u76F4\u3057\u3066\u304F\u3060\u3055\u3044\u3002
