--- spellchecker/hunspell-jmyspell/src/main/resources/org/omegat/spellchecker/Bundle_ru.properties
@@ -1,81 +1,0 @@
-#/**************************************************************************
-# OmegaT - Computer Assisted Translation (CAT) tool
-#          with fuzzy matching, translation memory, keyword search,
-#          glossaries, and translation leveraging into updated projects.
-#
-# Copyright (C) 2000-2006 Keith Godfrey, Maxym Mykhalchuk, and Henry Pijffers
-#               2007 Didier Briel, Zoltan Bartko, Martin Fleurke
-#               2008 Andrzej Sawula, Alex Buloichik, Didier Briel, Martin Fleurke
-#               2009 Didier Briel, Alex Buloichik, Martin Fleurke
-#               2010 Didier Briel, Wildrich Fourie, Martin Fleurke,
-#                    Antonio Vilei, Arno Peters, Alex Buloichik, Piotr Kulik
-#               2011 Didier Briel, Briac Pilpre
-#               2012 Martin Fleurke, Wildrich Fourie, Didier Briel, Guido Leenders, Jean-Christophe Helary,
-#                    Alex Buloichik
-#               2013 Aaron Madlon-Kay, Zoltan Bartko, Didier Briel
-#               2014 Aaron Madlon-Kay, Didier Briel, Piotr Kulik, Alex Buloichik, Briac Pilpre
-#               2015 Yu Tang, Aaron Madlon-Kay, Didier Briel, Hiroshi Miura
-#               2016 Alex Buloichik, Aaron Madlon-Kay, Didier Briel, Briac Pilpre
-#               2017 Didier Briel
-#               Home page: https://www.omegat.org/
-#               Support center: https://omegat.org/support
-#
-# This file is part of OmegaT.
-#
-# OmegaT is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# OmegaT is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#**************************************************************************/
-#
-#  Resources file for OmegaT
-#
-#  @author Keith Godfrey
-#  @author Maxym Mykhalchuk
-#  @author Henry Pijffers
-#  @author Didier Briel
-#  @author Jean-Christophe Helary
-#  @author Zoltan Bartko - bartkozoltan@bartkozoltan.com
-#  @author Martin Fleurke
-#  @author Andrzej Sawula
-#  @author Alex Buloichik
-#  @author Wildrich Fourie
-#  @author Antonio Vilei
-#  @author Briac Pilpre
-#  @author Aaron Madlon-Kay
-#  @author Piotr Kulik
-#  @author Briac Pilpre
-#
-#  @version 4.1
-
-
-
-GUI_SPELLCHECKER_TITLE=Параметры проверки орфографии
-PREFS_TITLE_SPELLCHECKER=Проверка орфографии
-GUI_SPELLCHECKER_AUTOSPELLCHECKCHECKBOX=&Автоматически проверять орфографию
-GUI_SPELLCHECKER_AVAILABLE_LABEL=Установленные словари:
-GUI_SPELLCHECKER_DICTIONARYLABEL=Папка словарей:
-GUI_SPELLCHECKER_FILE_CHOOSER_TITLE=Папка словарей
-GUI_SPELLCHECKER_DIRECTORYCHOOSERBUTTON=В&ыбрать…
-GUI_SPELLCHECKER_INSTALLBUTTON=Устан&овить…
-GUI_SPELLCHECKER_UNINSTALLBUTTON=&Удалить
-GUI_SPELLCHECKER_URL_LABEL=Ссылка на онлайн словари
-GUI_SPELLCHECKER_UNINSTALL_CURRENT=Будет удалён словарь проверки орфографии, который используется в текущем проекте. Продолжить?
-GUI_SPELLCHECKER_UNINSTALL_CURRENT_TITLE=Подтверждение удаления
-GUI_SPELLCHECKER_UNINSTALL_UNABLE=Не удалось удалить выбранный словарь
-GUI_SPELLCHECKER_UNINSTALL_UNABLE_TITLE=Ошибка удаления
-GUI_SPELLCHECKER_INSTALL_UNABLE_TITLE=Ошибка установки
-GUI_SPELLCHECKER_INSTALL_UNABLE=Сначала выберите папку для хранения словарей
-GUI_SPELLCHECKER_DIR_NOT_PRESENT_TITLE=Папка не существует
-GUI_SPELLCHECKER_DIR_NOT_PRESENT=Указанная папка для хранения словарей не существует. Хотите создать её?
-GUI_SPELLCHECKER_COULD_NOT_CREATE_DIR=Невозможно создать указанную папку.
-GUI_SPELLCHECKER_ERROR_ON_INSTALL=При скачивании словаря произошла ошибка. Пожалуйста, попробуйте снова.
-
