--- spellchecker/hunspell-jmyspell/src/main/resources/org/omegat/spellchecker/Bundle_zh_CN.properties
@@ -1,81 +1,0 @@
-#/**************************************************************************
-# OmegaT - Computer Assisted Translation (CAT) tool
-#          with fuzzy matching, translation memory, keyword search,
-#          glossaries, and translation leveraging into updated projects.
-#
-# Copyright (C) 2000-2006 Keith Godfrey, Maxym Mykhalchuk, and Henry Pijffers
-#               2007 Didi- Briel, Zoltan Bartko, Martin Fleurke
-#               2008 Andrzej Sawula, Alex Buloichik, Didier Briel, Martin Fleurke
-#               2009 Didier Briel, Alex Buloichik, Martin Fleurke
-#               2010 Didier Briel, Wildrich Fourie, Martin Fleurke,
-#                    Antonio Vilei, Arno Peters, Alex Buloichik, Piotr Kulik
-#               2011 Didier Briel, Briac Pilpre
-#               2012 Martin Fleurke, Wildrich Fourie, Didier Briel, Guido Leenders, Jean-Christophe Helary,
-#                    Alex Buloichik
-#               2013 Aaron Madlon-Kay, Zoltan Bartko, Didier Briel
-#               2014 Aaron Madlon-Kay, Didier Briel, Piotr Kulik, Alex Buloichik, Briac Pilpre
-#               2015 Yu Tang, Aaron Madlon-Kay, Didier Briel, Hiroshi Miura
-#               2016 Alex Buloichik, Aaron Madlon-Kay, Didier Briel, Briac Pilpre
-#               2017 Didier Briel
-#               2020-2022 Hiroshi Miura
-#               Home page: https://www.omegat.org/
-#               Support centre: https://omegat.org/support
-#
-# This file is part of OmegaT.
-#
-# OmegaT is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# OmegaT is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#**************************************************************************/
-#
-#  Resources file for OmegaT
-#
-#  @author Keith Godfrey
-#  @author Maxym Mykhalchuk
-#  @author Henry Pijffers
-#  @author Didier Briel
-#  @author Jean-Christophe Helary
-#  @author Zoltan Bartko - bartkozoltan@bartkozoltan.com
-#  @author Martin Fleurke
-#  @author Andrzej Sawula
-#  @author Alex Buloichik
-#  @author Wildrich Fourie
-#  @author Antonio Vilei
-#  @author Briac Pilpre
-#  @author Aaron Madlon-Kay
-#  @author Piotr Kulik
-#  @author Briac Pilpre
-#
-#  @version 4.1
-#
-
-GUI_SPELLCHECKER_TITLE=拼写检查器设置
-PREFS_TITLE_SPELLCHECKER=拼写检查器
-GUI_SPELLCHECKER_AUTOSPELLCHECKCHECKBOX=对译文进行拼写检查(&A)
-GUI_SPELLCHECKER_AVAILABLE_LABEL=可用语言：
-GUI_SPELLCHECKER_DICTIONARYLABEL=拼写检查词典位置：
-GUI_SPELLCHECKER_FILE_CHOOSER_TITLE=词典文件文件夹
-GUI_SPELLCHECKER_DIRECTORYCHOOSERBUTTON=浏览(&B)...
-GUI_SPELLCHECKER_INSTALLBUTTON=安装新语言(&I)...
-GUI_SPELLCHECKER_UNINSTALLBUTTON=删除(&E)
-GUI_SPELLCHECKER_URL_LABEL=可下载的拼写检查词典的 URL：
-GUI_SPELLCHECKER_UNINSTALL_CURRENT=您要删除当前项目正在使用的拼写检查词典。继续吗？
-GUI_SPELLCHECKER_UNINSTALL_CURRENT_TITLE=确认删除
-GUI_SPELLCHECKER_UNINSTALL_UNABLE=无法删除所选词典
-GUI_SPELLCHECKER_UNINSTALL_UNABLE_TITLE=删除失败
-GUI_SPELLCHECKER_INSTALL_UNABLE_TITLE=安装失败
-GUI_SPELLCHECKER_INSTALL_UNABLE=无法使用指定的拼写检查词典文件夹。
-GUI_SPELLCHECKER_DIR_NOT_PRESENT_TITLE=文件夹不存在
-GUI_SPELLCHECKER_DIR_NOT_PRESENT=指定的拼写检查词典文件夹不存在。要创建吗？
-GUI_SPELLCHECKER_COULD_NOT_CREATE_DIR=无法创建指定文件夹。
-GUI_SPELLCHECKER_ERROR_ON_INSTALL=下载拼写检查字典时发生错误。请稍后再试。
-
