+++ spellchecker/hunspell/src/main/resources/org/omegat/spellchecker/hunspell/Bundle_ja.properties
@@ -1,0 +1,26 @@
+#
+#  OmegaT - Computer Assisted Translation (CAT) tool
+#           with fuzzy matching, translation memory, keyword search,
+#           glossaries, and translation leveraging into updated projects.
+#
+#  Copyright (C) 2023 Hiroshi Miura
+#                Home page: https://www.omegat.org/
+#                Support center: https://omegat.org/support
+#
+#  This file is part of OmegaT.
+#
+#  OmegaT is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  OmegaT is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+SPELLCHECKER_HUNSPELL_INITIALIZED=\u8A00\u8A9E '{0}' \u7528\u306E LanguageTool Hunspell \u7DB4\u3044\u78BA\u8A8D\u7528\u8F9E\u66F8\u3092\u521D\u671F\u5316 {1} 
+SPELLCHECKER_HUNSPELL_EXCEPTION=hunspell\u306E\u8AAD\u307F\u8FBC\u307F\u30A8\u30E9\u30FC:{0}
