+++ spellchecker/hunspell/src/main/resources/org/omegat/spellchecker/hunspell/fr_FR.zip
