+++ spellchecker/hunspell/src/test/java/org/omegat/spellchecker/BundleTest.java
@@ -1,0 +1,51 @@
+/*
+ *  OmegaT - Computer Assisted Translation (CAT) tool
+ *           with fuzzy matching, translation memory, keyword search,
+ *           glossaries, and translation leveraging into updated projects.
+ *
+ *  Copyright (C) 2023 Hiroshi Miura
+ *                Home page: https://www.omegat.org/
+ *                Support center: https://omegat.org/support
+ *
+ *  This file is part of OmegaT.
+ *
+ *  OmegaT is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  OmegaT is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.omegat.spellchecker;
+
+import java.util.ResourceBundle;
+
+import org.junit.Test;
+
+import org.omegat.util.CommonVerifications;
+
+public class BundleTest extends CommonVerifications {
+    private static final String BUNDLE_BASENAME = "org.omegat.spellchecker.hunspell.Bundle";
+
+    @Test
+    public void testBundleEncodings() throws Exception {
+        assertBundle(BUNDLE_BASENAME);
+    }
+
+    @Test
+    public void testBundleLoading() {
+        assertBundleLoading(BUNDLE_BASENAME);
+    }
+
+    @Test
+    public void testUndefinedStrings() throws Exception {
+        assertBundleHasAllKeys(new String[]{"src"}, ResourceBundle.getBundle(BUNDLE_BASENAME));
+    }
+}
