+++ spellchecker/morfologik/plugin.properties
@@ -1,0 +1,31 @@
+#
+#  OmegaT - Computer Assisted Translation (CAT) tool
+#           with fuzzy matching, translation memory, keyword search,
+#           glossaries, and translation leveraging into updated projects.
+#
+#  Copyright (C) 2023 Hiroshi Miura
+#                Home page: https://www.omegat.org/
+#                Support center: https://omegat.org/support
+#
+#  This file is part of OmegaT.
+#
+#  OmegaT is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  OmegaT is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+License=GNU Public License version 3 or later
+Author=OmegaT team
+Version=6.0.0
+Category=spellchecker
+Name=Morfologik spell checker
+Class=org.omegat.spellchecker.morfologik.MorfologikSpellchecker
+Description=Morfologik spell checker module
